# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type HospitalType = typeof HospitalType[keyof typeof HospitalType];


export const HospitalType = {
  hospital: 'hospital',
  clinic: 'clinic',
} as const;

export interface Hospital {
  id: number;
  name: string;
  type: HospitalType;
  address: string;
  city: string;
  locality: string;
  contactPhone: string;
  /** @nullable */
  distanceKm: number | null;
  specialties: string[];
  doctorCount: number;
  rating: number;
  reviewCount: number;
  waitingMinutes: number;
  isOpen: boolean;
  isDemo: boolean;
  emergency: boolean;
  hours: string;
}

export interface Specialty {
  id: number;
  name: string;
  description: string;
}

export interface DashboardSummary {
  hospitalCount: number;
  specialtyCount: number;
  cityCount: number;
  cities: string[];
  featuredHospitals: Hospital[];
  specialties: Specialty[];
}

export interface Doctor {
  id: number;
  hospitalId: number;
  name: string;
  specialtyId: number;
  specialty: string;
  qualifications: string;
  yearsExperience: number;
  available: boolean;
}

export type ReviewStatus = typeof ReviewStatus[keyof typeof ReviewStatus];


export const ReviewStatus = {
  visible: 'visible',
  hidden: 'hidden',
} as const;

export interface Review {
  id: number;
  hospitalId: number;
  hospitalName: string;
  userId: string;
  userName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  text: string;
  status: ReviewStatus;
  createdAt: string;
  isDemo: boolean;
}

export type HospitalDetail = Hospital & {
  doctors: Doctor[];
  reviews: Review[];
};

export type HospitalInputType = typeof HospitalInputType[keyof typeof HospitalInputType];


export const HospitalInputType = {
  hospital: 'hospital',
  clinic: 'clinic',
} as const;

export interface HospitalInput {
  /** @minLength 1 */
  name: string;
  type: HospitalInputType;
  address: string;
  city: string;
  locality: string;
  contactPhone: string;
  /** @minimum 0 */
  waitingMinutes: number;
  isOpen: boolean;
  emergency: boolean;
  hours: string;
  specialtyIds: number[];
}

export type HospitalUpdateType = typeof HospitalUpdateType[keyof typeof HospitalUpdateType];


export const HospitalUpdateType = {
  hospital: 'hospital',
  clinic: 'clinic',
} as const;

export interface HospitalUpdate {
  /** @minLength 1 */
  name?: string;
  type?: HospitalUpdateType;
  address?: string;
  city?: string;
  locality?: string;
  contactPhone?: string;
  /** @minimum 0 */
  waitingMinutes?: number;
  isOpen?: boolean;
  emergency?: boolean;
  hours?: string;
  specialtyIds?: number[];
}

export interface SpecialtyInput {
  /** @minLength 1 */
  name: string;
  description: string;
}

export interface DoctorInput {
  hospitalId: number;
  /** @minLength 1 */
  name: string;
  specialtyId: number;
  qualifications: string;
  /** @minimum 0 */
  yearsExperience: number;
  available: boolean;
}

export interface DoctorUpdate {
  hospitalId?: number;
  /** @minLength 1 */
  name?: string;
  specialtyId?: number;
  qualifications?: string;
  /** @minimum 0 */
  yearsExperience?: number;
  available?: boolean;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /**
     * @minLength 1
     * @maxLength 1200
     */
  text: string;
}

export type ReviewModerationStatus = typeof ReviewModerationStatus[keyof typeof ReviewModerationStatus];


export const ReviewModerationStatus = {
  visible: 'visible',
  hidden: 'hidden',
} as const;

export interface ReviewModeration {
  status: ReviewModerationStatus;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt: string;
}

export interface ProfileUpdate {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}

export type AppointmentStatus = typeof AppointmentStatus[keyof typeof AppointmentStatus];


export const AppointmentStatus = {
  booked: 'booked',
  cancelled: 'cancelled',
} as const;

export interface Appointment {
  id: number;
  appointmentCode: string;
  patientName: string;
  hospitalId: number;
  hospitalName: string;
  specialtyId: number;
  specialty: string;
  doctorId: number;
  doctorName: string;
  appointmentDate: string;
  appointmentTime: string;
  status: AppointmentStatus;
  isDemo: boolean;
  createdAt: string;
}

export interface AppointmentInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  patientName: string;
  hospitalId: number;
  specialtyId: number;
  doctorId: number;
  appointmentDate: string;
  /** @minLength 1 */
  appointmentTime: string;
}

export type DemoTokenStatus = typeof DemoTokenStatus[keyof typeof DemoTokenStatus];


export const DemoTokenStatus = {
  active: 'active',
  completed: 'completed',
} as const;

export interface DemoToken {
  id: number;
  tokenNumber: string;
  hospitalId: number;
  hospitalName: string;
  specialtyId: number;
  specialty: string;
  waitingMinutes: number;
  status: DemoTokenStatus;
  isDemo: boolean;
  createdAt: string;
}

export interface TokenInput {
  hospitalId: number;
  specialtyId: number;
}

export interface VisitPlan {
  id: number;
  hospitalId: number;
  hospitalName: string;
  visitDate: string;
  purpose: string;
  notes: string;
  createdAt: string;
}

export interface VisitPlanInput {
  hospitalId: number;
  visitDate: string;
  /** @minLength 1 */
  purpose: string;
  notes: string;
}

export interface VisitPlanUpdate {
  hospitalId: number;
  visitDate: string;
  /** @minLength 1 */
  purpose: string;
  notes: string;
}

export interface EmergencyFacility {
  id: number;
  name: string;
  type: string;
  address: string;
  city: string;
  phone: string;
  category: string;
  availabilityNote: string;
  isDemo: boolean;
}

export type ListHospitalsParams = {
q?: string;
city?: string;
locality?: string;
type?: ListHospitalsType;
open?: boolean;
specialty?: string;
/**
 * @minimum -90
 * @maximum 90
 */
latitude?: number;
/**
 * @minimum -180
 * @maximum 180
 */
longitude?: number;
};

export type ListHospitalsType = typeof ListHospitalsType[keyof typeof ListHospitalsType];


export const ListHospitalsType = {
  hospital: 'hospital',
  clinic: 'clinic',
} as const;

export type ListEmergencyFacilitiesParams = {
city?: string;
category?: string;
};