# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [carepath]
      summary: Get the signed-in user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401": { description: Sign in required }
    patch:
      operationId: updateProfile
      tags: [carepath]
      summary: Update the signed-in user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileUpdate" }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserProfile" }
        "401": { description: Sign in required }
  /appointments:
    get:
      operationId: listAppointments
      tags: [carepath]
      summary: List the signed-in user's appointments
      responses:
        "200":
          description: User appointments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
        "401": { description: Sign in required }
    post:
      operationId: createAppointment
      tags: [carepath]
      summary: Book a prototype appointment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AppointmentInput" }
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "401": { description: Sign in required }
  /appointments/{id}/cancel:
    patch:
      operationId: cancelAppointment
      tags: [carepath]
      summary: Cancel an appointment owned by the signed-in user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Cancelled appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404": { description: Appointment not found }
        "401": { description: Sign in required }
  /tokens:
    get:
      operationId: listTokens
      tags: [carepath]
      summary: List the signed-in user's demo tokens
      responses:
        "200":
          description: User demo tokens
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/DemoToken" }
        "401": { description: Sign in required }
    post:
      operationId: createToken
      tags: [carepath]
      summary: Create a demo token for a hospital
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenInput" }
      responses:
        "201":
          description: Demo token created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemoToken" }
        "401": { description: Sign in required }
  /visit-plans:
    get:
      operationId: listVisitPlans
      tags: [carepath]
      summary: List the signed-in user's visit plans
      responses:
        "200":
          description: Visit plans
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/VisitPlan" }
        "401": { description: Sign in required }
    post:
      operationId: createVisitPlan
      tags: [carepath]
      summary: Create a visit plan
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VisitPlanInput" }
      responses:
        "201":
          description: Visit plan created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitPlan" }
        "401": { description: Sign in required }
  /visit-plans/{id}:
    patch:
      operationId: updateVisitPlan
      tags: [carepath]
      summary: Update a visit plan owned by the signed-in user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VisitPlanUpdate" }
      responses:
        "200":
          description: Updated visit plan
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitPlan" }
        "404": { description: Visit plan not found }
        "401": { description: Sign in required }
    delete:
      operationId: deleteVisitPlan
      tags: [carepath]
      summary: Delete a visit plan owned by the signed-in user
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Visit plan deleted }
        "404": { description: Visit plan not found }
        "401": { description: Sign in required }
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, admin] }
        createdAt: { type: string, format: date-time }
      required: [id, name, email, role, createdAt]
    ProfileUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
      required: [name]
    Appointment:
      type: object
      properties:
        id: { type: integer }
        appointmentCode: { type: string }
        patientName: { type: string }
        hospitalId: { type: integer }
        hospitalName: { type: string }
        specialtyId: { type: integer }
        specialty: { type: string }
        doctorId: { type: integer }
        doctorName: { type: string }
        appointmentDate: { type: string, format: date }
        appointmentTime: { type: string }
        status: { type: string, enum: [booked, cancelled] }
        isDemo: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, appointmentCode, patientName, hospitalId, hospitalName, specialtyId, specialty, doctorId, doctorName, appointmentDate, appointmentTime, status, isDemo, createdAt]
    AppointmentInput:
      type: object
      properties:
        patientName: { type: string, minLength: 1, maxLength: 120 }
        hospitalId: { type: integer }
        specialtyId: { type: integer }
        doctorId: { type: integer }
        appointmentDate: { type: string, format: date }
        appointmentTime: { type: string, minLength: 1 }
      required: [patientName, hospitalId, specialtyId, doctorId, appointmentDate, appointmentTime]
    DemoToken:
      type: object
      properties:
        id: { type: integer }
        tokenNumber: { type: string }
        hospitalId: { type: integer }
        hospitalName: { type: string }
        specialtyId: { type: integer }
        specialty: { type: string }
        waitingMinutes: { type: integer }
        status: { type: string, enum: [active, completed] }
        isDemo: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, tokenNumber, hospitalId, hospitalName, specialtyId, specialty, waitingMinutes, status, isDemo, createdAt]
    TokenInput:
      type: object
      properties:
        hospitalId: { type: integer }
        specialtyId: { type: integer }
      required: [hospitalId, specialtyId]
    VisitPlan:
      type: object
      properties:
        id: { type: integer }
        hospitalId: { type: integer }
        hospitalName: { type: string }
        visitDate: { type: string, format: date }
        purpose: { type: string }
        notes: { type: string }
        createdAt: { type: string, format: date-time }
      required: [id, hospitalId, hospitalName, visitDate, purpose, notes, createdAt]
    VisitPlanInput:
      type: object
      properties:
        hospitalId: { type: integer }
        visitDate: { type: string, format: date }
        purpose: { type: string, minLength: 1 }
        notes: { type: string }
      required: [hospitalId, visitDate, purpose, notes]
    VisitPlanUpdate:
      type: object
      properties:
        hospitalId: { type: integer }
        visitDate: { type: string, format: date }
        purpose: { type: string, minLength: 1 }
        notes: { type: string }
      required: [hospitalId, visitDate, purpose, notes]
...