# lib/api-zod/src/generated/types/tokenInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenInput {
  hospitalId: number;
  specialtyId: number;
}