# lib/api-zod/src/generated/types/visitPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitPlan {
  id: number;
  hospitalId: number;
  hospitalName: string;
  visitDate: Date;
  purpose: string;
  notes: string;
  createdAt: Date;
}