# lib/api-zod/src/generated/types/visitPlanInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CarePath healthcare discovery, planning, appointments, demo tokens, reviews, and administration API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitPlanInput {
  hospitalId: number;
  visitDate: Date;
  /** @minLength 1 */
  purpose: string;
  notes: string;
}